Extract OpenPathPrompt out of file_finder
