+++ crates/file_finder/Cargo.toml
@@ -21,6 +21,7 @@ futures.workspace = true
 fuzzy.workspace = true
 gpui.workspace = true
 menu.workspace = true
+open_path_prompt.workspace = true
 picker.workspace = true
 project.workspace = true
 schemars.workspace = true
