+++ crates/language_selector/Cargo.toml
@@ -15,11 +15,11 @@ doctest = false
 [dependencies]
 anyhow.workspace = true
 editor.workspace = true
-file_finder.workspace = true
 file_icons.workspace = true
 fuzzy.workspace = true
 gpui.workspace = true
 language.workspace = true
+open_path_prompt.workspace = true
 picker.workspace = true
 project.workspace = true
 settings.workspace = true
