+++ crates/open_path_prompt/Cargo.toml
@@ -1,22 +1,20 @@
 [package]
-name = "toolchain_selector"
+name = "open_path_prompt"
 version = "0.1.0"
-edition.workspace = true
 publish.workspace = true
+edition.workspace = true
 license = "GPL-3.0-or-later"
 
 [dependencies]
-anyhow.workspace = true
-convert_case.workspace = true
-editor.workspace = true
+file_icons.workspace = true
 futures.workspace = true
 fuzzy.workspace = true
 gpui.workspace = true
-language.workspace = true
-menu.workspace = true
-open_path_prompt.workspace = true
 picker.workspace = true
 project.workspace = true
+schemars.workspace = true
+serde.workspace = true
+settings.workspace = true
 ui.workspace = true
 util.workspace = true
 workspace.workspace = true
@@ -25,5 +23,4 @@ workspace.workspace = true
 workspace = true
 
 [lib]
-path = "src/toolchain_selector.rs"
-doctest = false
+path = "src/open_path_prompt.rs"
