+++ crates/open_path_prompt/LICENSE-GPL
@@ -1,0 +1,1 @@
+../../LICENSE-GPL
