+++ crates/open_path_prompt/src/file_finder_settings.rs
