+++ crates/open_path_prompt/src/open_path_prompt.rs
@@ -1,4 +1,9 @@
-use crate::file_finder_settings::FileFinderSettings;
+pub mod file_finder_settings;
+
+#[cfg(test)]
+mod open_path_prompt_tests;
+
+use file_finder_settings::FileFinderSettings;
 use file_icons::FileIcons;
 use futures::channel::oneshot;
 use fuzzy::{CharBag, StringMatch, StringMatchCandidate};
@@ -21,7 +26,7 @@ use util::{
 };
 use workspace::Workspace;
 
-pub(crate) struct OpenPathPrompt;
+pub struct OpenPathPrompt;
 
 pub struct OpenPathDelegate {
     tx: Option<oneshot::Sender<Option<Vec<PathBuf>>>>,
@@ -184,7 +189,7 @@ struct CandidateInfo {
 }
 
 impl OpenPathPrompt {
-    pub(crate) fn register(
+    pub fn register(
         workspace: &mut Workspace,
         _window: Option<&mut Window>,
         _: &mut Context<Workspace>,
@@ -196,7 +201,7 @@ impl OpenPathPrompt {
         }));
     }
 
-    pub(crate) fn register_new_path(
+    pub fn register_new_path(
         workspace: &mut Workspace,
         _window: Option<&mut Window>,
         _: &mut Context<Workspace>,
