+++ crates/recent_projects/Cargo.toml
@@ -24,7 +24,6 @@ db.workspace = true
 dev_container.workspace = true
 editor.workspace = true
 extension_host.workspace = true
-file_finder.workspace = true
 futures.workspace = true
 fuzzy.workspace = true
 gpui.workspace = true
@@ -33,6 +32,7 @@ log.workspace = true
 markdown.workspace = true
 menu.workspace = true
 node_runtime.workspace = true
+open_path_prompt.workspace = true
 ordered-float.workspace = true
 paths.workspace = true
 picker.workspace = true
