+++ crates/snippets_ui/Cargo.toml
@@ -12,11 +12,11 @@ workspace = true
 path = "src/snippets_ui.rs"
 
 [dependencies]
-file_finder.workspace = true
 file_icons.workspace = true
 fuzzy.workspace = true
 gpui.workspace = true
 language.workspace = true
+open_path_prompt.workspace = true
 paths.workspace = true
 picker.workspace = true
 settings.workspace = true
